+++ etc/GlobalSuppressions.Common.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) 2014, Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
+
+using System.Diagnostics.CodeAnalysis;
+
+#if !SIGNED_ASSEMBLY
+[assembly: SuppressMessage("Microsoft.Design", "CA2210:AssembliesShouldHaveValidStrongNames",
+    Justification = "Only released assemblies are signed.")]
+#endif