+++ src/Narvalo.Core/Properties/AssemblyInfo.cs
@@ -1,6 +1,11 @@
 ﻿// Copyright (c) 2014, Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
 
 using System.Reflection;
+using System.Runtime.CompilerServices;
 
 [assembly: AssemblyTitle("Narvalo.Core")]
 [assembly: AssemblyDescription("")]
+
+#if !NO_INTERNALS_VISIBLE_TO
+[assembly: InternalsVisibleTo("Narvalo.Facts" + Narvalo.Properties.AssemblyInfo.PublicKeySuffix)]
+#endif