+++ src/Narvalo.Mvp.Web/Properties/AssemblyInfo.cs
@@ -1,9 +1,14 @@
 ﻿// Copyright (c) 2014, Narvalo.Org. All rights reserved. See LICENSE.txt in the project root for license information.
 
 using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Narvalo.Mvp.Web")]
 [assembly: AssemblyDescription("")]
 
 [assembly: Guid("31740667-bd2d-4974-b0cf-0b7c0a639ff0")]
+
+#if !NO_INTERNALS_VISIBLE_TO
+[assembly: InternalsVisibleTo("Narvalo.Mvp.Facts" + Narvalo.Properties.AssemblyInfo.PublicKeySuffix)]
+#endif