+++ tests/Narvalo.Mvp.Facts/Resolvers/CompositeViewTypeResolverFacts.cs
@@ -22,6 +22,7 @@ namespace Narvalo.Mvp.Resolvers
 
         public static class ValidateViewTypeMethod
         {
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void ThrowsArgumentException_ForViewTypeOfClassType()
             {
@@ -31,7 +32,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Act & Assert
                 Assert.Throws<ArgumentException>(() => CompositeViewTypeResolver.ValidateViewType(viewType));
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void ThrowsArgumentException_ForViewTypeNotInheritingIView()
             {
@@ -41,7 +44,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Act & Assert
                 Assert.Throws<ArgumentException>(() => CompositeViewTypeResolver.ValidateViewType(viewType));
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void ThrowsArgumentException_ForViewTypeOfPrivateType()
             {
@@ -51,7 +56,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Act & Assert
                 Assert.Throws<ArgumentException>(() => CompositeViewTypeResolver.ValidateViewType(viewType));
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void ThrowsArgumentException_ForViewTypeContainingPublicMethods()
             {
@@ -61,7 +68,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Act & Assert
                 Assert.Throws<ArgumentException>(() => CompositeViewTypeResolver.ValidateViewType(viewType));
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void Passes_ForViewTypeOfIViewType()
             {
@@ -74,7 +83,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Assert
                 Assert.True(true);
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void Passes_ForViewTypeInheritingIView()
             {
@@ -87,7 +98,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Assert
                 Assert.True(true);
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void Passes_ForViewTypeInheritingGenericIView()
             {
@@ -100,7 +113,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Assert
                 Assert.True(true);
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void Passes_ForViewTypeOfGenericIViewType()
             {
@@ -113,7 +128,9 @@ namespace Narvalo.Mvp.Resolvers
                 // Assert
                 Assert.True(true);
             }
+#endif
 
+#if NO_INTERNALS_VISIBLE_TO
             [Fact]
             public static void Passes_ForViewTypeContainingPropertiesAndEventHandlers()
             {
@@ -126,6 +143,7 @@ namespace Narvalo.Mvp.Resolvers
                 // Assert
                 Assert.True(true);
             }
+#endif
         }
 
         #region Helper classes